requirements.txt
pyotp==2.9.0
qrcode[pil]==7.4.2
pillow==10.0.1
pyperclip==1.9.0
